Added migration scripts.  Dropping lb_usage_event and event_type table (and reverse for rollback.)
Removed last vestiges of lb_usage and lb_usage_event tables in the loadbalancing_usage database.  Added migrations to rename or remove unused tables in the loadbalancing database.

Removed ip_version as a column in the AccessList entity.  Removed all usages of it.  Added migration to remove it from table.